+++ Celeste_Launcher_Gui/Forms/GameScan.cs
@@ -23,8 +23,10 @@ namespace Celeste_Launcher_Gui.Forms
             else
                 tb_GamePath.Text = GameScannnerApi.GetGameFilesRootPath();
 
-            if (Program.UserConfig != null && Program.UserConfig.BetaUpdate)
-                checkBox1.Checked = true;
+            if (Program.UserConfig != null && Program.UserConfig.IsLegacyXLive)
+                rB_Legacy.Checked = true;
+            else
+                rB_Sparta.Checked = true;
         }
         
         private void BtnRunScan_Click(object sender, EventArgs e)
@@ -32,10 +34,10 @@ namespace Celeste_Launcher_Gui.Forms
             try
             {
                 Program.UserConfig.GameFilesPath = tb_GamePath.Text;
-                Program.UserConfig.BetaUpdate = checkBox1.Checked;
+                Program.UserConfig.IsLegacyXLive = rB_Legacy.Checked;
                 Program.UserConfig.Save(Program.UserConfigFilePath);
 
-                using (var form = new GameScanProgressForm(Program.UserConfig.GameFilesPath, Program.UserConfig.IsSteamVersion, Program.UserConfig.BetaUpdate))
+                using (var form = new GameScanProgressForm(Program.UserConfig.GameFilesPath, Program.UserConfig.IsSteamVersion, Program.UserConfig.IsLegacyXLive))
                 {
                     var dr = form.ShowDialog();
 
@@ -79,6 +81,16 @@ namespace Celeste_Launcher_Gui.Forms
             {
                 //
             }
+        }
+
+        private void RB_Sparta_CheckedChanged(object sender, EventArgs e)
+        {
+            Program.UserConfig.IsLegacyXLive = !rB_Sparta.Checked;
+        }
+
+        private void RB_Legacy_CheckedChanged(object sender, EventArgs e)
+        {
+            Program.UserConfig.IsLegacyXLive = rB_Legacy.Checked;
         }
     }
 }